chart

+++ app/src/main/java/com/example/Business_Converter/CoursesCalculator/models/SingleRate.kt
@@ -1,8 +1,10 @@
 package com.example.Business_Converter.CoursesCalculator.models
 
+import java.math.BigDecimal
+
 
 data class SingleFinalRate(
     val effectiveDate: String,
-    val mid: Double,
+    val mid: BigDecimal,
     val no: String
 )
